% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_crs.R, R/records.R
\name{as.data.frame,extent_crs-method}
\alias{as.data.frame,extent_crs-method}
\alias{as.data.frame,extent_crs}
\alias{as.data.frame,records-method}
\title{Coerce to a Data Frame}
\usage{
\S4method{as.data.frame}{extent_crs}(x)

\S4method{as.data.frame}{records}(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
returns a data frame, normally with all row names
}
\description{
Functions to check if an object is a data frame, or coerce it if possible.
}
\examples{
# load example rtoi
file.copy(from=system.file("ex/Navarre",package="rsat"),
         to=tempdir(),
         recursive = TRUE)

navarre <- read_rtoi(file.path(tempdir(),"Navarre"))

# get the records
rcds <- records(navarre)
# coerce the records to rtoi
df <- as.data.frame(rcds)
# print the dataframe
print(df)
}
