% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{as.records}
\alias{as.records}
\alias{as.records,data.frame-method}
\alias{as.records,data.frame}
\title{Creates records object from data frame}
\usage{
as.records(x)

\S4method{as.records}{data.frame}(x)
}
\arguments{
\item{x}{data frame with columns representing the slots of records.}
}
\description{
Creates records object from data frame
}
