% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_crs.R, R/records.R
\name{c,extent_crs-method}
\alias{c,extent_crs-method}
\alias{c}
\alias{c,records-method}
\title{Combine values into a vector or a list}
\usage{
\S4method{c}{extent_crs}(x, ...)

\S4method{c}{records}(x, ...)
}
\arguments{
\item{x}{a \code{records} object.}

\item{...}{additional arguments.}
}
\value{
a combination of 'x' class elements
}
\description{
This is a generic function which combines its arguments.
}
\details{
The default method combines its arguments to form a vector.
All arguments are coerced to a common type which is the type
 of the returned value. All attributes except names are removed.
}
