% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_mask.R
\name{cloud_mask}
\alias{cloud_mask}
\alias{cloud_mask,rtoi-method}
\alias{cloud_mask,rtoi}
\title{Create cloud mask from an rtoi}
\usage{
cloud_mask(x, ...)

\S4method{cloud_mask}{rtoi}(x, products = "ALL", verbose = FALSE, overwrite = FALSE, ...)
}
\arguments{
\item{x}{rtoi object from which cloud masks are computed.}

\item{...}{additional arguments}

\item{products}{the name of the dataset from which cloud masks are computed.}

\item{verbose}{logical argument. If \code{TRUE}, the function prints the
running steps and warnings.}

\item{overwrite}{logical argument. If \code{TRUE}, overwrites the existing
images with the same name.}
}
\description{
Create cloud mask from an rtoi
}
