% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R, R/rtoi.R
\name{dates}
\alias{dates}
\alias{dates,records-method}
\alias{dates,records}
\alias{dates<-}
\alias{dates<-,records}
\alias{dates<-,records-method}
\alias{dates,rtoi-method}
\alias{dates,rtoi}
\title{Gets the dates from a records or an rtoi}
\usage{
dates(x)

\S4method{dates}{records}(x)

dates(x) <- value

\S4method{dates}{records}(x) <- value

\S4method{dates}{rtoi}(x)
}
\arguments{
\item{x}{a records or an rtoi object.}

\item{value}{date argument. The new value for x.}
}
\description{
Gets the dates from a records or an rtoi
}
