% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex.madrid}
\alias{ex.madrid}
\title{A polygon with the border of Madrid (Spain)}
\description{
Spatial feature (\code{sf}) representing the border of Madrid with
coordinates in the longitude/latitude format.
}
\keyword{data}
