% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex.manhattan}
\alias{ex.manhattan}
\title{A polygon with the border of Manhattan (USA)}
\description{
Spatial feature (\code{sf}) representing the border of Manhattan with
coordinates in the NAD83 format.
}
\keyword{data}
