% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{get_api_name}
\alias{get_api_name}
\alias{get_api_name,records-method}
\alias{get_api_name,records}
\title{Gets the api name of the records}
\usage{
get_api_name(x)

\S4method{get_api_name}{records}(x)
}
\arguments{
\item{x}{records object.}
}
\value{
a character vector containing the api name of al the elements in x.
}
\description{
Functions to get or set the names of an object.
}
