% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoi.R
\name{get_database}
\alias{get_database}
\alias{get_database,rtoi-method}
\alias{get_database,rtoi}
\title{Extract database path}
\usage{
get_database(x)

\S4method{get_database}{rtoi}(x)
}
\arguments{
\item{x}{rtoi object}
}
\description{
get_database returns a character with the database of an rtoi.
}
