% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R, R/rtoi.R
\name{get_dir}
\alias{get_dir}
\alias{get_dir,records-method}
\alias{get_dir,records}
\alias{get_order,records-method}
\alias{get_dir,rtoi-method}
\alias{get_dir,rtoi}
\alias{get_dir<-}
\alias{get_dir,rtoi,character}
\title{get the path from records or rtoi}
\usage{
get_dir(x)

\S4method{get_dir}{records}(x)

\S4method{get_order}{records}(x)

\S4method{get_dir}{rtoi}(x)

get_dir(x) <- value
}
\arguments{
\item{x}{records/rtoi object}

\item{value}{character argument. The value for change the direcory of x.}
}
\description{
get the path from records or rtoi
}
