% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{get_download}
\alias{get_download}
\title{Extract download url}
\usage{
get_download(x)
}
\arguments{
\item{x}{records object}
}
\description{
get_download returns a vector of character with the url for download the image.
}
