% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{get_order}
\alias{get_order}
\alias{get_order<-}
\alias{get_order<-,records-method}
\alias{get_order<-,records}
\title{get the order slot from records or rtoi}
\usage{
get_order(x)

get_order(x) <- value

\S4method{get_order}{records}(x) <- value
}
\arguments{
\item{x}{records/rtoi object}

\item{value}{logical argument. The value for change the value of x.}
}
\description{
get the order slot from records or rtoi
}
