% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{get_preview}
\alias{get_preview}
\alias{get_preview,records-method}
\alias{get_preview,records}
\alias{get_download,records-method}
\title{Extract the url of the preview}
\usage{
get_preview(x)

\S4method{get_preview}{records}(x)

\S4method{get_download}{records}(x)
}
\arguments{
\item{x}{a \code{records} object.}
}
\value{
preview url of a records
}
\description{
It returns a character vector of urls to preview the data records.
}
\examples{
# load example rtoi
navarre <- read_rtoi(system.file("ex/Navarre",package="rsat"))

# get the records
rcds <- records(navarre)

# get a vector with the preview url of each record
get_api_name(rcds)
}
