% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoi.R
\name{get_raster}
\alias{get_raster}
\alias{get_raster,rtoi-method}
\alias{get_raster,rtoi}
\title{get the raster of a variable from rtoi object}
\usage{
get_raster(x, p, v, ...)

\S4method{get_raster}{rtoi}(x, p, v, ...)
}
\arguments{
\item{x}{the target rtoi}

\item{p}{character argument. The product name to be extracted.}

\item{v}{character argument. The derived variable to be extracted.}

\item{...}{additional arguments.}
}
\value{
a raster stack with the variable and product defined by the user.
}
\description{
get the raster of a variable from rtoi object
}
\examples{
\dontrun{
library(rsat)

# load navarre sf from the package
data(ex.navarre)

# set the credentials
set_credentials("username","password")

# path where the region is stored
rtoi.path <- tempdir()
# path where downloads are stored
db.path <- file.path(tempdir(),"DATABASE")
navarre<-new_rtoi("Navarre",
                  ex.navarre,
                  rtoi.path,
                  db.path)#'
#Landsat-5
sat_search(region=navarre,
          product="LANDSAT_TM_C1",
          dates=as.Date("1988-08-01")+seq(1,35))
download(navarre)

mosaic(navarre,overwrite=T)

derive(navarre,"NDVI",product="LANDSAT_TM_C1")
ls6.ndvi<-get_raster(navarre,"LANDSAT_TM_C1","NDVI")
}
}
