% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaic.R
\name{mosaic,rtoi,ANY-method}
\alias{mosaic,rtoi,ANY-method}
\alias{mosaic,records,ANY-method}
\alias{mosaic,records}
\title{Mosaics the tiles covering the region of interest}
\usage{
\S4method{mosaic}{rtoi,ANY}(x, y, ...)

\S4method{mosaic}{records,ANY}(
  x,
  out_path,
  db_path,
  bfilter,
  warp = "extent",
  region,
  overwrite = FALSE,
  ...
)
}
\arguments{
\item{x}{an rtoi object with a records that cover the region of interest.}

\item{y}{Omitted}

\item{...}{additional arguments}

\item{db_path}{path where the images will be stored when using an rtoi.
By default, the path is defined by the rtoi.}

\item{bfilter}{a vector with the bands to be extracted when \code{untar=TRUE}. If not supplied, all are extracted.}

\item{warp}{character argument. Defines how to warp the resulting tile.}

\item{region}{a \code{Spatial*}, projected \code{raster*}, o
\code{sf} class object defining the region of interest.}

\item{out.dir}{path where the images will be stored when using a records.}
}
\description{
Mosaics the tiles covering the region of interest
}
\examples{
\dontrun{
library(rsat)

# load navarre sf from the package
data(ex.navarre)

# set the credentials
set_credentials("username","password")

# path where the region is stored
rtoi.path <- tempdir()
# path where downloads are stored
db.path <- file.path(tempdir(),"DATABASE")
navarre<-new_rtoi("Navarre",
                  ex.navarre,
                  rtoi.path,
                  db.path)#'
#Landsat-5
sat_search(region=navarre,
          product="LANDSAT_TM_C1",
          dates=as.Date("1988-08-01")+seq(1,35))
download(navarre)

mosaic(navarre,overwrite=T)

derive(navarre,"NDVI",product="LANDSAT_TM_C1")
}
}
