% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R, R/rtoi.R
\name{names,records-method}
\alias{names,records-method}
\alias{names,rtoi-method}
\alias{names,rtoi}
\alias{names<-,rtoi,character-method}
\alias{names<-,rtoi,character}
\title{Gets the name of the object}
\usage{
\S4method{names}{records}(x)

\S4method{names}{rtoi}(x)

\S4method{names}{rtoi,character}(x) <- value
}
\arguments{
\item{x}{a records or an rtoi object.}

\item{value}{character argument. The new value for x.}
}
\value{
a character vector containing the name of all the names in x.
}
\description{
Functions to get or set the names of an object.
}
\examples{
\dontrun{
data(ex.navarre)
# path where the data will be
rtoi.path <- tempdir()
# path where downloads are stored
db.path <- file.path(tempdir(),"DATABASE")
navarre<-new_rtoi("Navarre",
                  ex.navarre,
                  rtoi.path,
                  db.path)
names(navarre)

# Create a set of records using sat_search function
s2.lvl2.result<-sat_search(region=ex.navarre,
                          product="S2MSI2A",
                          dates=as.Date("2019-01-01")+seq(1,30,1))
names(s2.lvl2.result)

names(s2.lvl2.result)<-"New name"
names(s2.lvl2.result)
}
}
