% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{new_record}
\alias{new_record}
\alias{new_record,character,character,Date,character,character,character,numeric,numeric,character,character,character,logical,extent_crs-method}
\alias{new_record,character,character,Date,character,character,character,numeric,numeric,character,character,character,logical,extent_crs}
\alias{new_record,character,character,Date,character,character,character,numeric,numeric,character,character,character,logical,missing-method}
\alias{new_record,character,character,Date,character,character,character,numeric,numeric,character,character,character,logical,missing}
\title{Creates a new records}
\usage{
new_record(
  sat,
  name,
  date,
  product,
  download,
  file_path,
  path,
  row,
  tileid,
  preview,
  api_name,
  order,
  extent_crs
)

\S4method{new_record}{character,character,Date,character,character,character,numeric,numeric,character,character,character,logical,extent_crs}(
  sat,
  name,
  date,
  product,
  download,
  file_path,
  path,
  row,
  tileid,
  preview,
  api_name,
  order,
  extent_crs
)

\S4method{new_record}{character,character,Date,character,character,character,numeric,numeric,character,character,character,logical,missing}(
  sat,
  name,
  date,
  product,
  download,
  file_path,
  path,
  row,
  tileid,
  preview,
  api_name,
  order
)
}
\arguments{
\item{sat}{the name of the satellite which the record belongs}

\item{name}{the name of the record}

\item{date}{the date of the record}

\item{product}{the product}

\item{download}{the download path of the record}

\item{file_path}{the directory path where the image will be download}

\item{path}{the path of the tile sistem}

\item{row}{the path of the tile sistem}

\item{tileid}{the id of the tile}

\item{preview}{the url of the preview image of the record}

\item{api_name}{the api name}

\item{order}{boolean slot, defines if the image need to be order or not}

\item{extent_crs}{extent information for projecting and preview the slot}
}
\value{
records object
}
\description{
Creates a records object from scrach
}
\examples{
\dontrun{
a<-new_record(sat="modis",
name="mod09a",
date=as.Date("2011087","\%Y\%j"),
product="product",
download="url/aaa/download")
}
}
