% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,rtoi,Date-method}
\alias{plot,rtoi,Date-method}
\alias{plot,rtoi,character-method}
\alias{plot,character}
\alias{plot,records,ANY-method}
\alias{plot,records}
\alias{plot,rtoi,missing-method}
\alias{plot,rtoi,missing}
\title{Plot an \code{rtoi} object}
\usage{
\S4method{plot}{rtoi,Date}(
  x,
  y,
  ...,
  variable = "rgb",
  band_name = c("red", "green", "blue"),
  verbose = FALSE,
  xsize = 250,
  ysize = 250
)

\S4method{plot}{rtoi,character}(
  x,
  y,
  ...,
  variable = "rgb",
  product = "ALL",
  dates = NULL,
  verbose = FALSE,
  xsize = 250,
  ysize = 250
)

\S4method{plot}{records,ANY}(x, y, verbose = FALSE, ...)

\S4method{plot}{rtoi,missing}(x, y, verbose = FALSE, ...)
}
\arguments{
\item{x}{an \code{rtoi} or \code{records}.}

\item{y}{character argument. The valid values are "dates", "preview", or "view".}

\item{...}{additional arguments.}

\item{variable}{character argument. The variable to be plotted. By default, a color (RGB) variable is selected .}

\item{band_name}{character vector argument. Enables false color plots. By default, usual bands are selected \code{c("red","green","blue")}.}

\item{verbose}{logical argument. If \code{TRUE}, the function prints the
running steps and warnings.}

\item{xsize}{the number of samples on the horizontal axis.}

\item{ysize}{the number of samples on the vertical axis.}

\item{product}{character argument. The product name to be plotted.}

\item{dates}{date vector argument. The dates to be plotted.}
}
\value{
\code{tmap} plot.
}
\description{
Plot (a map of) the values of an \code{rtoi} or \code{records} object.
}
\examples{
\dontrun{
# plotting records
a<-5+5

}
}
