% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview}
\alias{preview}
\alias{preview,rtoi,Date-method}
\alias{preview,rtoi,date}
\alias{preview,rtoi,missing-method}
\alias{preview,rtoi,missing}
\alias{preview,records,Date-method}
\alias{preview,records,date}
\alias{preview,records,numeric-method}
\alias{preview,rtoi,numeric}
\title{Previews a records or an rtoi object}
\usage{
preview(x, n, ...)

\S4method{preview}{rtoi,Date}(x, n, lpos = c(3, 2, 1), add.layer = FALSE, verbose = FALSE, ...)

\S4method{preview}{rtoi,missing}(x, n, lpos = c(3, 2, 1), add.layer = FALSE, verbose = FALSE, ...)

\S4method{preview}{records,Date}(
  x,
  n,
  lpos = c(3, 2, 1),
  tmp_dir = file.path(tempdir()),
  add.layer = FALSE,
  verbose = FALSE,
  get.map = TRUE,
  ...
)

\S4method{preview}{records,numeric}(
  x,
  n,
  lpos = c(3, 2, 1),
  tmp_dir = file.path(tempdir()),
  add.layer = FALSE,
  verbose = FALSE,
  get.map = TRUE,
  ...
)
}
\arguments{
\item{x}{a records or an rtoi object.}

\item{n}{the date expressed as the temporal index in the time series.}

\item{...}{additional arguments}

\item{lpos}{vector argument. Defines the position of the red-green-blue
layers to enable false color visualization.}

\item{add.layer}{logical argument. If \code{TRUE}, the function plots the
image on an existing map.}

\item{verbose}{logical argument. If \code{TRUE}, the function prints the
running steps and warnings.}

\item{tmp_dir}{character argument. The directory where previsualization images
will be located.}

\item{get.map}{logical argument. If \code{TRUE}, the function return the leaflet map.}
}
\description{
Previews a records or an rtoi object
}
\examples{
\dontrun{
library(rsat)

# load navarre sf from the package
data(ex.navarre)

# set the credentials
set_credentials("username","password")

# path where the region is stored
rtoi.path <- tempdir()
# path where downloads are stored
db.path <- file.path(tempdir(),"DATABASE")
navarre<-new_rtoi("Navarre",
                  ex.navarre,
                  rtoi.path,
                  db.path)#'

sat_search(region=navarre,
          product="S2MSI1C",
          dates=as.Date("2018-01-01")+seq(1,30,1))

preview(navarre)


sat_search(region=ex.navarre,
          product="S2MSI1C",
          dates=as.Date("2018-01-01")+seq(1,30,1))

preview(navarre,n=1)
}
}
