% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R, R/extent_crs.R, R/records.R, R/rtoi.R,
%   R/variables.R
\name{print,api-method}
\alias{print,api-method}
\alias{print,api}
\alias{print,extent_crs-method}
\alias{print,extent_crs}
\alias{print,records-method}
\alias{print,records}
\alias{print,rtoi-method}
\alias{print,variables-method}
\title{Print Values}
\usage{
\S4method{print}{api}(x)

\S4method{print}{extent_crs}(x)

\S4method{print}{records}(x)

\S4method{print}{rtoi}(x)

\S4method{print}{variables}(x, ...)
}
\arguments{
\item{x}{print object.}

\item{...}{additional arguments.}
}
\description{
print prints its argument and returns it invisibly (via invisible(x)). It is a generic function which means that new printing methods can be easily added for new classes.
}
\examples{
\dontrun{
data(ex.navarre)
# path where the data will be
rtoi.path <- tempdir()
# path where downloads are stored
db.path <- file.path(tempdir(),"DATABASE")
navarre<-new_rtoi("Navarre",
                  ex.navarre,
                  rtoi.path,
                  db.path)
print(navarre)
}
}
