% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{print_credentials}
\alias{print_credentials}
\alias{print_credentials,ANY-method}
\title{prints the credentials defined by the user}
\usage{
print_credentials(...)

\S4method{print_credentials}{ANY}()
}
\arguments{
\item{...}{additional arguments.}
}
\description{
prints the credentials defined by the user
}
\examples{
print_credentials()
set_credentials("example","example","earthdata")
print_credentials()
}
