% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R, R/rtoi.R
\name{product}
\alias{product}
\alias{product,records-method}
\alias{product,records}
\alias{product,rtoi-method}
\alias{product,rtoi}
\title{get the product name from records or rtoi}
\usage{
product(x)

\S4method{product}{records}(x)

\S4method{product}{rtoi}(x)
}
\arguments{
\item{x}{records/rtoi object}
}
\description{
get the product name from records or rtoi
}
