% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoi.R
\name{read_rtoi}
\alias{read_rtoi}
\alias{read_rtoi,character-method}
\alias{read_rtoi,character}
\title{Reads an rtoi from the hard drive}
\usage{
read_rtoi(path, ...)

\S4method{read_rtoi}{character}(path, ...)
}
\arguments{
\item{path}{an rtoi object.}

\item{...}{additional arguments.}
}
\description{
Reads an rtoi from the hard drive
}
\examples{
\dontrun{
#' data(ex.navarre)
# path where the data will be
rtoi.path <- tempdir()
# path where downloads are stored
db.path <- file.path(tempdir(),"DATABASE")
navarre<-new_rtoi("Navarre",
                  ex.navarre,
                  rtoi.path,
                  db.path)
rm(navarre)

rtoi.path <- file.path(tempdir(),"Navarre")
new.navarre <- read_rtoi(rtoi.path)
print(new.navarre)
}
}
