% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\docType{class}
\name{records-class}
\alias{records-class}
\title{A class object for satellite image metadata}
\description{
This class sorts the main attributes of satellite image metadata
to ease previewing, downloading, and managing data.
}
\details{
\code{records} works as vector with multiple parameters, and
accepts usual R methods such as \code{c}, \code{[]},
\code{length}, \code{subset} or \code{unique}.

The object can be coerced into a \code{data.frame} by
using the command \code{as.data.frame}.
}
\section{Slots}{

\describe{
\item{\code{sat}}{the name of the satellite to which the record belongs.}

\item{\code{name}}{the name of the satellite record.}

\item{\code{date}}{the date of the satellite record.}

\item{\code{product}}{the product.}

\item{\code{path}}{the path of the tile system.}

\item{\code{row}}{the row of the tile system.}

\item{\code{tileid}}{the tile id.}

\item{\code{download}}{the url to download the satellite record.}

\item{\code{file_path}}{the saving directory for the satellite record.}

\item{\code{preview}}{the url of the preview of the satellite record.}

\item{\code{api_name}}{the api name.}

\item{\code{order}}{boolean, defines if the image must be requested or not.}

\item{\code{extent_crs}}{extent (used to project the preview).}
}}

\examples{
\dontrun{
data(ex.navarre)
# Create a records of Sentinel-2 Level 1C images
s2.lvl1.result<-sat_search(region=ex.navarre,
                           product="S2MSI1C",
                           dates=as.Date("2018-01-01")+seq(1,30,1))

# Create a records of Sentinel-2 Level 2A images
s2.lvl2.result<-sat_search(region=ex.navarre,
                          product="S2MSI2A",
                          dates=as.Date("2019-01-01")+seq(1,30,1))
class(s2.lvl2.result)
dates(s2.lvl2.result)
all.records <- c(s2.lvl1.result,s2.lvl2.result)

print(all.records)
}
}
