% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoi.R
\name{records}
\alias{records}
\alias{records,rtoi-method}
\alias{records,rtoi}
\alias{records<-}
\alias{records<-,rtoi,records-method}
\alias{records<-,rtoi,records}
\title{Extracts the satellite records}
\usage{
records(x)

\S4method{records}{rtoi}(x)

records(x) <- value

\S4method{records}{rtoi,records}(x) <- value
}
\arguments{
\item{x}{an rtoi object}

\item{value}{a records object to be set to x.}
}
\description{
returns the object records from an rtoi.
}
