% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoi.R
\name{region}
\alias{region}
\alias{region,rtoi-method}
\alias{region,rtoi}
\alias{region<-}
\alias{region<-,rtoi,sf-method}
\alias{region<-,rtoi,sf}
\title{Extract region}
\usage{
region(x)

\S4method{region}{rtoi}(x)

region(x) <- value

\S4method{region}{rtoi,sf}(x) <- value
}
\arguments{
\item{x}{rtoi object}

\item{value}{sf argument. The value for change the region of x.}
}
\description{
region returns an sf with the region of an rtoi.
}
