% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoi.R
\name{rename}
\alias{rename}
\alias{rename,rtoi,character-method}
\alias{rename,rtoi,character}
\title{Renames an \code{rtoi}}
\usage{
rename(x, newname)

\S4method{rename}{rtoi,character}(x, newname)
}
\arguments{
\item{x}{an rtoi object}

\item{newname}{a character class to rename the \code{rtoi}.}
}
\description{
Renames all parameters and folder name of an \code{rtoi}.
}
\examples{
\dontrun{
myrtoi <- read_rtoi("file_path/rtoir_name")
rename(myrtoi,"Navarre_BACK")
}
}
