% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{rsat_products}
\alias{rsat_products}
\alias{rsat_products,ANY-method}
\title{Show the products accepted by the services}
\usage{
rsat_products(...)

\S4method{rsat_products}{ANY}()
}
\arguments{
\item{...}{additional arguments.}
}
\value{
prints a list of products
}
\description{
Show the products accepted by the services
}
\examples{
rsat_products()
}
