% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoi.R
\docType{class}
\name{rtoi-class}
\alias{rtoi-class}
\title{Region Time Of Interest (rtoi)}
\description{
It is a proxy object for managing an spatial region in large time periods
of time and with data from multiple satellites. The rtoi stores all the metadata
of a region of interest to process the region.
}
\details{
An \code{rtoi} object manages two main folders, the database and the rtoi folder.
When \code{download} function is called with an rtoi, all the images will be
stored in the database.
The rtoi folder includes a \code{region.rtoi} file with all the metadata. Once
the \code{mosaic} function is run, multiple folders with the name of the satellites
considered in the rtoi apears. Each folder contais a preproceses version of
each satellite.
}
\section{Fields}{

\describe{
\item{\code{name}}{character field. Takes the name of the region of interest}

\item{\code{rtoi_path}}{character field. The path where the rtoi will be stored.}

\item{\code{region}}{sf field. An sf with the region of interest}

\item{\code{records}}{records field. The records that presents the images for your
region and time of interest.}

\item{\code{db_path}}{character field. The directory where the original version
of the images are stored.}
}}


\examples{
\dontrun{
data(ex.navarre)

# path where the region is stored
rtoi.path <- tempdir()

# path where downloads are stored
db.path <- file.path(tempdir(),"DATABASE")
navarre<-new_rtoi("Navarre",
                  ex.navarre,
                  rtoi.path,
                  db.path)

print(navarre)

navarre.records <- records(navarre)

print(navarre.records)
}
}
