% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{set_credentials}
\alias{set_credentials}
\alias{set_credentials,character,character,missing-method}
\alias{set_credentials,character,character,missing}
\alias{set_credentials,character,character,character-method}
\alias{set_credentials,character,character,character}
\title{asigns the credentials for data searching and downloading}
\usage{
set_credentials(user, pass, credential)

\S4method{set_credentials}{character,character,missing}(user, pass)

\S4method{set_credentials}{character,character,character}(user, pass, credential)
}
\arguments{
\item{user}{character argument. Defines the username of an api platform
for searching and downloading images}

\item{pass}{character argument. Defines the password of an api platform
for searching and downloading images}

\item{credential}{optional argument. Defines the platform name. earthdata,
scihub, scihubs5p, or ALL are valid credentials}
}
\description{
asigns the credentials for data searching and downloading
}
\examples{
print_credentials()
set_credentials("example","example")
print_credentials()
set_credentials("example","example","earthdata")
print_credentials()
}
