% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoi.R
\name{set_database}
\alias{set_database}
\alias{set_database,rtoi-method}
\alias{set_database,rtoi}
\title{#' @export
setGeneric("get_database<-",function(x,value)  standardGeneric("get_database<-"))
setMethod("get_database<-",
          signature(x = "rtoi",value="character"),
          function(x,value){
            x$db_path<-value
            write_rtoi(x)
            return(x)
          })
set database path}
\usage{
set_database(x, value)

\S4method{set_database}{rtoi}(x, value)
}
\arguments{
\item{x}{rtoi object}

\item{value}{character argument. The value for change the database directory of x.}
}
\description{
set_database set a new database  directory in x.
}
