% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{show_variables}
\alias{show_variables}
\alias{show_variables,ANY-method}
\alias{show_variables-method}
\title{List the variables and satellites supported by \code{rsat}}
\usage{
show_variables(...)

\S4method{show_variables}{ANY}()
}
\arguments{
\item{...}{arguments for nestering functions}
}
\value{
prints supported satellites and derived variables information.
}
\description{
Displays the satellites and variable method
}
\examples{
show_variables()
}
