% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{[,records,ANY,ANY,ANY-method}
\alias{[,records,ANY,ANY,ANY-method}
\title{Extracts or replace parts of an object}
\usage{
\S4method{[}{records,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s)}

\item{i}{numeric argument. The the position of the element to select/modify.}

\item{value}{records argument. The slot of the records to be changed.}
}
\description{
Operators acting on vectors, matrices, arrays and lists to extract or replace parts.
}
