% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{subset,records-method}
\alias{subset,records-method}
\title{get the order slot from records or rtoi}
\usage{
\S4method{subset}{records}(x, subset, select)
}
\arguments{
\item{x}{records/rtoi object}

\item{subset}{R object with the value for subsetting}

\item{select}{character argument indicating the name of the slot to use in subsetting.}
}
\description{
get the order slot from records or rtoi
}
