% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing_function.R
\name{test_function}
\alias{test_function}
\title{Testing function}
\usage{
test_function()
}
\description{
Function used for testing some internal functions in continuous integration.
}
\examples{
test_function()
}
