% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{unique,records,ANY-method}
\alias{unique,records,ANY-method}
\title{Extract unique elements}
\usage{
\S4method{unique}{records,ANY}(x)
}
\arguments{
\item{x}{a \code{records} object.}
}
\description{
It returns a \code{records} like \code{x} but with duplicate
elements/rows removed.
}
