% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{unique,records,ANY-method}
\alias{unique,records,ANY-method}
\alias{unique}
\title{Extract unique elements}
\usage{
\S4method{unique}{records,ANY}(x)
}
\arguments{
\item{x}{a \code{records} object.}
}
\value{
unique elements in records class
}
\description{
It returns a \code{records} like \code{x} but with duplicate
elements/rows removed.
}
\examples{
\dontrun{
# load example rtoi
file.copy(from=system.file("ex/Navarre",package="rsat"),
         to=tempdir(),
         recursive = TRUE)

navarre <- read_rtoi(file.path(tempdir(),"Navarre"))

# get the records
rcds <- records(navarre)

duplicate.records <- c(rcds[1],rcds[1])
length(duplicate.records)
print(duplicate.records)
single.record <- unique(duplicate.records)
length(single.record)
print(single.record)
}
}
