% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R, R/rtoi.R
\name{dates}
\alias{dates}
\alias{dates,records-method}
\alias{dates<-}
\alias{dates<-,records-method}
\alias{dates,rtoi-method}
\title{Get the dates from a \code{records} or an \code{rtoi}}
\usage{
dates(x)

\S4method{dates}{records}(x)

dates(x) <- value

\S4method{dates}{records}(x) <- value

\S4method{dates}{rtoi}(x)
}
\arguments{
\item{x}{a \code{records} or an \code{rtoi} object.}
}
\value{
returns a vector of \code{Date} class
}
\description{
Get the dates from a \code{records} or an \code{rtoi}
}
