% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R, R/rtoi.R
\name{dates}
\alias{dates}
\alias{dates,records-method}
\alias{dates,records}
\alias{dates<-}
\alias{dates<-,records}
\alias{dates<-,records-method}
\alias{dates,rtoi-method}
\alias{dates,rtoi}
\title{get date name from records or rtoi}
\usage{
dates(x)

\S4method{dates}{records}(x)

dates(x) <- value

\S4method{dates}{records}(x) <- value

\S4method{dates}{rtoi}(x)
}
\arguments{
\item{x}{records/rtoi object}

\item{value}{date argument. The value for change the date o x.}
}
\description{
get date name from records or rtoi
}
