% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive}
\alias{derive}
\alias{derive,rtoi,character-method}
\alias{derive,rtoi,character}
\title{Compute a remote sensing index from a rtoi}
\usage{
derive(x, variable, ...)

\S4method{derive}{rtoi,character}(
  x,
  variable,
  product,
  dates,
  fun,
  overwrite = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{rtoi object for deriving variables}

\item{variable}{the name of the variable, run "show_variables()"
to check the variables supported by the package}

\item{...}{additional argument for variable deriving}

\item{product}{the name of the dataset for deriving variables.}

\item{dates}{a vector with the dates being considered
for creating cloud masks. This argument is optional.}

\item{fun}{a \code{function} that computes the remote sensing index.}

\item{overwrite}{logical argument. If \code{TRUE}, overwrites the existing
images with the same name.}

\item{verbose}{logical argument. If \code{TRUE}, the function prints the
running steps and warnings.}
}
\value{
nothing. The derived variables will be save in the hard drive.
Use get_stars to get the variables.
}
\description{
Compute a remote sensing index from a rtoi
}
\examples{
\dontrun{
library(rsat)

# load navarre sf from the package
data(ex.navarre)

# set the credentials
set_credentials("username","password")

# path where the region is stored
rtoi.path <- tempdir()
# path where downloads are stored
db.path <- file.path(tempdir(),"DATABASE")
navarre<-new_rtoi("Navarre",
                  ex.navarre,
                  rtoi.path,
                  db.path)#'
#Landsat-5
sat_search(region=navarre,
          product="LANDSAT_TM_C1",
          dates=as.Date("1988-08-01")+seq(1,35))
download(navarre)

mosaic(navarre,overwrite=T)

derive(navarre,"NDVI",product="LANDSAT_TM_C1")
}
}
