% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\alias{download,rtoi-method}
\alias{download,rtoi}
\alias{download,records-method}
\alias{download,records}
\title{Download the images from a \code{records} or an \code{rtoi} object}
\usage{
download(x, ...)

\S4method{download}{rtoi}(x, db_path, ...)

\S4method{download}{records}(x, out.dir, verbose = FALSE, ...)
}
\arguments{
\item{x}{a \code{records} or an \code{rtoi} object.}

\item{...}{additional arguments.}

\item{db_path}{path to the database. By default, the path
is defined by the \code{rtoi}.}

\item{out.dir}{path where the outputs are stored when using a \code{records}.}

\item{verbose}{logical argument. If \code{TRUE}, the function prints the
running steps and warnings.}
}
\description{
The function saves the raw images in the database or the specified directory.
It skips the images that already exist in the database or directory.
}
\examples{
\dontrun{
library(rsat)

# load navarre sf from the package
data(ex.navarre)

# set the credentials
set_credentials("username", "password")

# path where the region is stored
rtoi.path <- tempdir()
# path where downloads are stored
db.path <- file.path(tempdir(), "DATABASE")
navarre <- new_rtoi(
  "Navarre",
  ex.navarre,
  rtoi.path,
  db.path
)

# Landsat-5
sat_search(
  region = navarre,
  product = "LANDSAT_TM_C1",
  dates = as.Date("1988-08-01") + seq(1, 35)
)
download(navarre)
}
}
