% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoi.R
\name{get_database}
\alias{get_database}
\alias{get_database,rtoi-method}
\alias{get_database,rtoi}
\alias{get_database,missing-method}
\title{Extracts the path to the database}
\usage{
get_database(x)

\S4method{get_database}{rtoi}(x)

\S4method{get_database}{missing}()
}
\arguments{
\item{x}{an rtoi object.}
}
\description{
extracts the path to the database from an rtoi.
}
