% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{get_download}
\alias{get_download}
\title{Extract the url to download a data record}
\usage{
get_download(x)
}
\arguments{
\item{x}{a \code{records} object.}
}
\description{
It returns a character with the url to download the image.
}
