% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{get_download}
\alias{get_download}
\title{Extracts the url to download a record}
\usage{
get_download(x)
}
\arguments{
\item{x}{a records object}
}
\description{
get_download returns a character with the url to download the image.
}
