% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{get_order}
\alias{get_order}
\alias{get_order<-}
\alias{get_order<-,records-method}
\alias{get_order<-,records}
\title{Gets the slot called order from a records or an rtoi}
\usage{
get_order(x)

get_order(x) <- value

\S4method{get_order}{records}(x) <- value
}
\arguments{
\item{x}{a records or an rtoi object.}

\item{value}{logical argument. The new value for x.}
}
\description{
Gets the slot called order from a records or an rtoi
}
