% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{get_order}
\alias{get_order}
\alias{get_order<-}
\alias{get_order<-,records-method}
\alias{get_order<-,records}
\title{Get the slot called order from a \code{records} or an \code{rtoi}}
\usage{
get_order(x)

get_order(x) <- value

\S4method{get_order}{records}(x) <- value
}
\arguments{
\item{x}{a \code{records} or an \code{rtoi} object.}

\item{value}{logical argument. The new value for \code{x}.}
}
\description{
Get the slot called order from a \code{records} or an \code{rtoi}
}
\examples{
# load example rtoi
navarre <- read_rtoi(system.file("ex/Navarre",package="rsat"))

# get the records
rcds <- records(navarre)

# gets a boolean
get_order(rcds)
}
