% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{get_preview}
\alias{get_preview}
\alias{get_preview,records-method}
\alias{get_preview,records}
\alias{get_download,records-method}
\title{Extract preview url}
\usage{
get_preview(x)

\S4method{get_preview}{records}(x)

\S4method{get_download}{records}(x)
}
\arguments{
\item{x}{records object}
}
\description{
get_preview returns a vector of character with the url for previewing the recods.
}
