% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoi.R
\name{new_rtoi}
\alias{new_rtoi}
\alias{new_rtoi,character,sf,character,character,missing-method}
\alias{new_rtoi,character,sf,character,character,missing}
\alias{new_rtoi,character,sf,character,character,records-method}
\alias{character,sf,character,character,records}
\title{Creates a new rtoi object}
\usage{
new_rtoi(name, sfobj, rtoi_path, db_path, records)

\S4method{new_rtoi}{character,sf,character,character,missing}(name, sfobj, rtoi_path, db_path)

\S4method{new_rtoi}{character,sf,character,character,records}(name, sfobj, rtoi_path, db_path, records)
}
\arguments{
\item{name}{the name of the region of interes}

\item{sfobj}{sf object.}

\item{rtoi_path}{the directory where the rtoi will be located.}

\item{db_path}{the directory where the database for creating the rtoi will be located.}

\item{records}{records object.}
}
\value{
the reference of the rtoi object
}
\description{
Creates a new rtoi object
}
