% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,rtoi,ANY-method}
\alias{plot,rtoi,ANY-method}
\alias{plot,records,ANY-method}
\alias{plot,records}
\alias{plot,rtoi,missing-method}
\alias{plot,rtoi,missing}
\title{Plots an rtoi object}
\usage{
\S4method{plot}{rtoi,ANY}(x, y, verbose = FALSE, xsize = 250, ysize = 250, ...)

\S4method{plot}{records,ANY}(x, y, verbose = FALSE, ...)

\S4method{plot}{rtoi,missing}(x, y, verbose = FALSE, ...)
}
\arguments{
\item{x}{an rtoi}

\item{y}{character argument. The product name to be plotted.}

\item{verbose}{logical argument. If \code{TRUE}, the function prints the
running steps and warnings.}

\item{xsize}{the size of x dimension for subsetting.}

\item{ysize}{the size of y dimension for subsetting.}

\item{...}{additional arguments.}
}
\value{
tmap plot.
}
\description{
Plot (a map of) the values of an rtoi object.


}
\details{

}
