% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R, R/rtoi.R
\name{product}
\alias{product}
\alias{product,records-method}
\alias{product,records}
\alias{product,rtoi-method}
\alias{product,rtoi}
\title{Gets the name of the product from a records or an rtoi}
\usage{
product(x)

\S4method{product}{records}(x)

\S4method{product}{rtoi}(x)
}
\arguments{
\item{x}{a records or an rtoi object.}
}
\description{
Gets the name of the product from a records or an rtoi
}
