% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R, R/rtoi.R
\name{product}
\alias{product}
\alias{product,records-method}
\alias{product,records}
\alias{product,rtoi-method}
\alias{product,rtoi}
\title{Get the name of the product from a \code{records} or an \code{rtoi}}
\usage{
product(x)

\S4method{product}{records}(x)

\S4method{product}{rtoi}(x)
}
\arguments{
\item{x}{a \code{records} or an \code{rtoi} object.}
}
\value{
the name of the product in the records
}
\description{
Get the name of the product from a \code{records} or an \code{rtoi}
}
