% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\docType{class}
\name{records-class}
\alias{records-class}
\title{Satellite image metadata in R class}
\description{
This class sorts the main satellite image metadata attributes
to ease the image preview, download, and data management.
}
\details{
\code{records} works as vector with multiple parameters, and
accepts usuals R methods used in R as \code{c}, \code{[]},
\code{length}, \code{subset} or \code{unique}.

All the data in the class can be substracted as \code{dataframe}
using the method \code{as.data.frame}.
}
\section{Slots}{

\describe{
\item{\code{sat}}{the name of the satellite which the record belongs}

\item{\code{name}}{the name of the record}

\item{\code{date}}{the date of the record}

\item{\code{product}}{the product}

\item{\code{path}}{the path of the tile sistem}

\item{\code{row}}{the path of the tile sistem}

\item{\code{tileid}}{the id of the tile}

\item{\code{download}}{the download path of the record}

\item{\code{file_path}}{the directory path where the image will be download}

\item{\code{preview}}{the url of the preview image of the record}

\item{\code{api_name}}{the api name}

\item{\code{order}}{boolean slot, defines if the image need to be order or not}

\item{\code{extent_crs}}{extent information for projecting and preview the slot}
}}

\examples{
\dontrun{
data(ex.navarre)
# Create a set of records using sat_search function
s2.lvl1.result<-sat_search(region=ex.navarre,
                           product="S2MSI1C",
                           dates=as.Date("2018-01-01")+seq(1,30,1))

# Create a set of records using sat_search function
s2.lvl2.result<-sat_search(region=ex.navarre,
                          product="S2MSI2A",
                          dates=as.Date("2019-01-01")+seq(1,30,1))
class(s2.lvl2.result)
dates(s2.lvl2.result)
all.records <- c(s2.lvl1.result,s2.lvl2.result)

print(all.records)
}
}
