% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoi.R
\name{records}
\alias{records}
\alias{records,rtoi-method}
\alias{records,rtoi}
\alias{records<-}
\alias{records<-,rtoi,records-method}
\alias{records<-,rtoi,records}
\title{Extract records}
\usage{
records(x)

\S4method{records}{rtoi}(x)

records(x) <- value

\S4method{records}{rtoi,records}(x) <- value
}
\arguments{
\item{x}{rtoi object}

\item{value}{records argument. The value for change the records of x.}
}
\description{
records returns an records object from an rtoi.
}
