% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoi.R
\name{region}
\alias{region}
\alias{region,rtoi-method}
\alias{region,rtoi}
\alias{region<-}
\alias{region<-,rtoi,sf-method}
\alias{region<-,rtoi,sf}
\title{Extracts region from an rtoi}
\usage{
region(x)

\S4method{region}{rtoi}(x)

region(x) <- value

\S4method{region}{rtoi,sf}(x) <- value
}
\arguments{
\item{x}{an rtoi object.}

\item{value}{an sf object to define the region in x.}
}
\description{
gets the sf that specifies the region of an rtoi.
}
