% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoi.R
\docType{class}
\name{rtoi-class}
\alias{rtoi-class}
\title{Region and Time Of Interest (rtoi)}
\description{
It is a proxy object to store the metadata about the satellite imagery
that covers a spatial region over a time period. Images can come from
multiple satellite programs and its purpose is to help managing
heterogeneous datasets.
}
\details{
An \code{rtoi} object manages two main folders called database and rtoi.
The database is meant to work as a local, generic, and organized archive
of raw satellite images. When the function \code{download} is called
with an rtoi, images go to the database.
In contrast, the rtoi folder contains just the relevant information for
a particular region and time of interest. It contains a \code{region.rtoi}
file which saves all the metadata about that region/time of interest
and satellite imagery available. Also, when the function \code{mosaic} is
run, the relevant images from the database are cropped, mosaicked, and saved
in the rtoi folder.
}
\section{Fields}{

\describe{
\item{\code{name}}{a character with the name of the region of interest}

\item{\code{rtoi_path}}{a character with the path to the rtoi folder.}

\item{\code{region}}{an sf with the region of interest.}

\item{\code{records}}{the satellite records available for your region and time of interest.}

\item{\code{db_path}}{a character with the path to the database.}
}}


\examples{
\dontrun{
data(ex.navarre)

# path where the region is stored
rtoi.path <- tempdir()

# path where downloads are stored
db.path <- file.path(tempdir(),"DATABASE")
navarre<-new_rtoi(name="Navarre",
                  region=ex.navarre,
                  rtoi_path=rtoi.path,
                  db_path=db.path)

print(navarre)

navarre.records <- records(navarre)

print(navarre.records)
}
}
