% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R, R/rtoi.R
\name{sat_name}
\alias{sat_name}
\alias{sat_name,records-method}
\alias{sat_name,records}
\alias{sat_name,rtoi-method}
\alias{sat_name,rtoi}
\title{get satellite name from records or rtoi}
\usage{
sat_name(x)

\S4method{sat_name}{records}(x)

\S4method{sat_name}{rtoi}(x)
}
\arguments{
\item{x}{records/rtoi object}
}
\description{
get satellite name from records or rtoi
}
