% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoi.R
\name{set_database}
\alias{set_database}
\alias{set_database,rtoi-method}
\alias{set_database,rtoi}
\alias{set_database,character-method}
\title{set database path}
\usage{
set_database(x, ...)

\S4method{set_database}{rtoi}(x, value)

\S4method{set_database}{character}(x)
}
\arguments{
\item{x}{rtoi object}

\item{value}{character argument. The value for
change the database directory of x.}
}
\description{
set_database set a new database  directory in x.
}
