% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_crs.R, R/records.R, R/rtoi.R,
%   R/variables.R
\name{show,extent_crs-method}
\alias{show,extent_crs-method}
\alias{show,extent_crs}
\alias{show,records-method}
\alias{show,rtoi-method}
\alias{show,rtoi}
\alias{show,variables-method}
\alias{show,records}
\title{Show an object}
\usage{
\S4method{show}{extent_crs}(object)

\S4method{show}{records}(object)

\S4method{show}{rtoi}(object)

\S4method{show}{variables}(object)
}
\arguments{
\item{object}{Any R Object}
}
\description{
Display the object, by printing, plotting or whatever suits its class.
This function has specialized methods. Use \code{showDefault()} to call
the default method.
}
\details{
Formal methods for the function \code{show} will usually be invoked
automatically (see the details).
}
