% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent_crs.R, R/records.R
\name{[,extent_crs,ANY,ANY-method}
\alias{[,extent_crs,ANY,ANY-method}
\alias{sub,extent_crs}
\alias{[<-,extent_crs,ANY,ANY-method}
\alias{sub<-,extent_crs}
\alias{[,records,ANY,ANY-method}
\alias{[<-,records,ANY,ANY-method}
\alias{'[<-',rtoi,records}
\title{Extracts or replace parts of an object}
\usage{
\S4method{[}{extent_crs,ANY,ANY}(x, i)

\S4method{[}{extent_crs,ANY,ANY}(x, i) <- value

\S4method{[}{records,ANY,ANY}(x, i)

\S4method{[}{records,ANY,ANY}(x, i) <- value
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s)}

\item{i}{numeric argument. The the position of the element to select/modify.}

\item{value}{records argument. The slot of the records to be changed.}
}
\description{
Operators acting on vectors, matrices, arrays and lists to extract or replace parts.
}
