% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{subset,records-method}
\alias{subset,records-method}
\title{Filters the satellite records of a records or an rtoi object}
\usage{
\S4method{subset}{records}(x, subset, select)
}
\arguments{
\item{x}{a records or an rtoi object.}

\item{subset}{an R object with the value for subsetting.}

\item{select}{character argument indicating the name of the slot.}
}
\description{
Filters the satellite records of a records or an rtoi object
}
