% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{unique,records,ANY-method}
\alias{unique,records,ANY-method}
\title{Extract Unique Elements}
\usage{
\S4method{unique}{records,ANY}(x)
}
\arguments{
\item{x}{records object}
}
\description{
unique returns records like x but with duplicate elements/rows removed.
}
